fixed mistype of a condition

+++ file_io/1-create_file.c
@@ -25,7 +25,7 @@ int create_file(const char *filename, char *text_content)
 		while (text_content[len])
 		len++;
 		Write = write(fd, text_content, len);
-		if (Write == -1 || Write != len)
+		if (Write == -1 || (size_t)Write != len)
 		{
 			close(fd);
 			return (-1);
